add all controllers and validations

+++ src/controller/authorController.js
@@ -1,12 +1,11 @@
 //Recogemos express
 const express = require('express')
-//Recogemos el ROUTER para poder crear los metodos GET, POST ...
+//Creamos un objeto ROUTER (y le ponemos el nombre authorRouter) donde guardaremos las rutas del controlador con los metodos GET, POST ...
 const authorRouter = express.Router()
-//Llamamos a nuestro fichero de servicios para poder usar las funciones
-//en nuestros metodos GET, POST ... (que está en model/authors)
+//Importamos el servicio entero (que contendrá los métodos) de autores, para poder usar los métodos
 const authorService = require('../model/authors')
 //Llamamos al fichero donde hacemos las validaciones, para poder usarlas en los metodos
-const bookValidationService = require('../services/bookValidationService')
+const authorValidationService = require('../services/authorValidationService')
 //Llamamos a ajv para poder validar esquemas
 const Ajv = require('ajv');
 const ajv = new Ajv();
@@ -14,9 +13,26 @@ const ajv = new Ajv();
 
 
 
+//POST Global
+authorRouter.post('', async (req, res) => {
+    const { data } = req.body
+    //Validacion del esquema
+    let valid = authorValidationService.validateAuthor(authorValidationService.authorCreateSchema, data)
+    //
+    if(valid){
+        const id = await authorService.createAuthor(data)
+        res.status(201).json({message: `Author with id ${id /*data.id*/} was created `})
+    }
+    else{
+        res.status(400).json({message: `Author schema is not valid`})
+    }
+      
+})
+
+
 //GET global
-authorRouter.get('', (req, res) => {
-    const authors = authorService.consultAllAuthors()
+authorRouter.get('', async (req, res) => {
+    const authors = await authorService.consultAllAuthors()
     if(authors.length > 0){
         res.status(200).json({authors})    
     }
@@ -26,11 +42,10 @@ authorRouter.get('', (req, res) => {
 })
 
 //GET por id
-authorRouter.get('/:id', (req, res) => {
+authorRouter.get('/:id', async (req, res) => {
     const { id } = req.params;
-    const authors = authorService.consultAuthorsById(id)
+    const authors = await authorService.consultAuthorsById(id)
     if(!authors) {
-        
         res.status(404).json({ message: `Author with id ${id} was not found `});
     }
     else {
@@ -39,9 +54,9 @@ authorRouter.get('/:id', (req, res) => {
 })
 
 //GET por name
-authorRouter.get('/name/:name', (req, res) => {
+authorRouter.get('/name/:name', async (req, res) => {
     const { name } = req.params;
-    const authors = authorService.consultAuthorsByName(name)
+    const authors = await authorService.consultAuthorsByName(name)
     if(!authors) {
         res.status(404).json({ message: `Author with name ${name} was not found `});
     }
@@ -50,37 +65,75 @@ authorRouter.get('/name/:name', (req, res) => {
     }
 })
 
-//POST Global
-authorRouter.post('', (req, res) => {
-    const { data } = req.body
-    authorService.createAuthor(data)
-    res.status(201).json({message: `Author with id ${data.id} was created `})
-})
-
 //PUT by id (Con el put modificamos campos de una entiedad, pero tenemos que enviar todos)
-authorRouter.put('/:id', (req, res) => {
+authorRouter.put('/:id', async (req, res) => {
     const { data } = req.body
     const { id } = req.params
-    console.log(data)
-    console.log(id)
-    authorService.updateAuthor(data, id)
-    res.status(200).json({ message: `Author with id ${id} was updated` })
+    //Primero validamos si el autor que queremos updatear existe. Si existe lo updateamos y si no, no.
+    const exist = await authorValidationService.entityExists('authors', id)
+    //
+    if(exist){
+        //Si existe validamos que el body trae lo necesario para hacer el PUT
+        let valid = authorValidationService.validateAuthor(authorValidationService.authorUpdateSchemaPut, data)
+        //
+        if(valid){
+            await authorService.updateAuthor(data, id)
+            res.status(200).json({ message: `Author with id ${id} was updated` })
+        }
+        else{
+            res.status(400).json({ message: `Author update schema is not valid` })
+        }
+    }
+    else{
+        res.status(404).json({ message: `Author with id ${id} was not founded` })
+    }
 })
+    
 
 //PATCH by id (Con el patch modificamos campos de una entidad, y podemos solo mandar los campos modificados)
-authorRouter.patch('/:id', (req, res) => {
+authorRouter.patch('/:id', async (req, res) => {
     const { data } = req.body
     const { id } = req.params
-    authorService.updateAuthor(data, id)
-    res.status(200).json({ message: `Author with id ${id} was updated` })
+    // Primero validamos si el autor que queremos updatear existe. Si existe lo updateamos y si no, no.
+    const exist = await authorValidationService.entityExists('authors', id)
+    //
+    if(exist){
+        //Si existe validamos que el body trae lo necesario para hacer el PATCH     
+        const valid = authorValidationService.validateAuthor(authorValidationService.authorUpdateSchemaPatch, data)
+        //
+        if(valid){
+            await authorService.updateAuthor(data, id)
+            res.status(200).json({ message: `Author with id ${id} was updated` })
+        }
+        else{
+            res.status(400).json({ message: `Author update schema is not valid` })
+        }
+    }
+    else{
+        res.status(404).json({ message: `Author with id ${id} was not founded` })
+    }
+    
+    
 })
 
+
 //DELETE by id
-authorRouter.delete('/:id', (req, res) => {
+authorRouter.delete('/:id', async (req, res) => {
     const { id } = req.params
-    authorService.deleteAuthor(id)
-    res.status(200).json({ message: `Author with id ${id} was deleted` })
+    //Primero validamos si el autor que queremos borrar existe. Si existe lo borramos y si no, no.
+    const exist = await authorValidationService.entityExists('authors', id)
+    //
+    if(exist){
+        await authorService.deleteAuthor(id)
+        res.status(200).json({ message: `Author with id ${id} was deleted` })
+    }
+    else{
+        res.status(404).json({ message: `Author with id ${id} was not founded` })
+    }
+    
+    
 })
+
 
 
 //Exportamos el ROUTER
